Fix shader and seperate gpu from presentation
