+++ index.html
@@ -4,6 +4,8 @@
 <head>
   <meta charset="UTF-8" />
   <link rel="icon" type="image/ico" href="/favicon.ico" />
+  <link rel="stylesheet" href="/src/css/style.css">
+  <link rel="stylesheet" href="/src/css/ui-styles.css">
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Kayo</title>
 </head>
