+++ src/ts/app.ts
@@ -1,2 +1,4 @@
+import { loop } from "./rendering/tempRender";
 import { rootSplitPaneContainer } from "./ui/ui";
 document.body.appendChild(rootSplitPaneContainer);
+requestAnimationFrame(loop);
