+++ src/ts/rendering/Config.ts
@@ -1,5 +1,9 @@
+import { gpu } from "./gpuInit";
+
+export type SwapChainFormat = "rgba16float" | "bgra8unorm" | "rgba8unorm";
+
 export default class {
-	public static swapChainFormat: GPUTextureFormat & ("rgba16float" | "bgra8unorm" | "rgba8unorm") = "rgba16float";
-	public static swapChainColorSpace: PredefinedColorSpace = "display-p3";
-	public static swapChainToneMappingMode: GPUCanvasToneMappingMode = "extended";
+	public static swapChainFormat: SwapChainFormat = gpu.getPreferredCanvasFormat() as SwapChainFormat;
+	public static swapChainColorSpace: PredefinedColorSpace = "srgb";
+	public static swapChainToneMappingMode: GPUCanvasToneMappingMode = "standard";
 }
