+++ src/wgsl/shader.wgsl
@@ -20,29 +20,65 @@ fn vertex_main2(
 		return VertexOut(vec4f(position.x * 0.5, position.y * 0.5, position.zw), vec4f(0));
 }
 
-fn linearRGBToDisplayP3(linearRGB: vec3<f32>) -> vec3<f32> {
-    // Transformation matrix from linear RGB (sRGB) to Display P3
+fn applyGammaCorrection(linearRGB: vec3<f32>) -> vec3<f32> {
+    var gammaCorrectedRGB: vec3<f32>;
+
+    if (linearRGB.x <= 0.0031308) {
+        gammaCorrectedRGB.x = 12.92 * linearRGB.x;
+    } else {
+        gammaCorrectedRGB.x = 1.055 * pow(linearRGB.x, 1.0 / 2.4) - 0.055;
+    }
+
+    if (linearRGB.y <= 0.0031308) {
+        gammaCorrectedRGB.y = 12.92 * linearRGB.y;
+    } else {
+        gammaCorrectedRGB.y = 1.055 * pow(linearRGB.y, 1.0 / 2.4) - 0.055;
+    }
+
+    if (linearRGB.z <= 0.0031308) {
+        gammaCorrectedRGB.z = 12.92 * linearRGB.z;
+    } else {
+        gammaCorrectedRGB.z = 1.055 * pow(linearRGB.z, 1.0 / 2.4) - 0.055;
+    }
+
+    return gammaCorrectedRGB;
+}
+
+fn revertGammaCorrection(gammaCorrectedRGB: vec3<f32>) -> vec3<f32> {
+    var linearRGB: vec3<f32>;
+
+    if (gammaCorrectedRGB.x <= 0.04045) {
+        linearRGB.x = gammaCorrectedRGB.x / 12.92;
+    } else {
+        linearRGB.x = pow((gammaCorrectedRGB.x + 0.055) / 1.055, 2.4);
+    }
+
+    if (gammaCorrectedRGB.y <= 0.04045) {
+        linearRGB.y = gammaCorrectedRGB.y / 12.92;
+    } else {
+        linearRGB.y = pow((gammaCorrectedRGB.y + 0.055) / 1.055, 2.4);
+    }
+
+    if (gammaCorrectedRGB.z <= 0.04045) {
+        linearRGB.z = gammaCorrectedRGB.z / 12.92;
+    } else {
+        linearRGB.z = pow((gammaCorrectedRGB.z + 0.055) / 1.055, 2.4);
+    }
+
+    return linearRGB;
+}
+
+fn linearRGBToLinearDisplayP3(linearRGB: vec3<f32>) -> vec3<f32> {
     let transformationMatrix = mat3x3<f32>(
         0.875905, 0.035332, 0.016382,
         0.122070, 0.964542, 0.063767,
         0.002025, 0.000126, 0.919851
     );
-
-    // Apply the transformation matrix
-    var p3Color = transformationMatrix * linearRGB;
-
-    // Convert linear Display P3 to gamma-corrected Display P3
-    let threshold = vec3<f32>(0.0031308);
-    let lowValue = 12.92 * p3Color;
-    let highValue = 1.055 * pow(p3Color, vec3<f32>(1.0 / 2.4)) - vec3<f32>(0.055);
-    
-    // Blend between lowValue and highValue based on the condition
-    p3Color = mix(lowValue, highValue, vec3<f32>(p3Color > threshold));
-
-    return p3Color;
+    return transformationMatrix * linearRGB;
 }
 
 @fragment
 fn fragment_main(vertexData: VertexOut) -> @location(0) vec4f {
-		return vec4f(linearRGBToDisplayP3(vertexData.color.xyz), 1);
+		return vec4f(applyGammaCorrection(linearRGBToLinearDisplayP3(revertGammaCorrection(vertexData.color.xyz) * 1.5)), 1);
+        // return vertexData.color * 2;
 }
